+++ virtual_machine/vm_sti.c
@@ -33,9 +33,9 @@ void	load_res(t_vm *vm, t_cmd *cmd, int direct)
 int		vm_rdd_sti(t_vm *vm, t_cmd *cmd)
 {
 	//int		pc;
-	int		res;
-	short	arg1;
-	short	arg2;
+	int				res;
+	unsigned short	arg1;
+	unsigned short	arg2;
 	
 	res = 0;
 	//pc = 0xFF & (vm->arena[mdx(cmd->idx + 2)].acb);
@@ -49,23 +49,23 @@ int		vm_rdd_sti(t_vm *vm, t_cmd *cmd)
 	arg2 <<= 8;
 	arg2 += 0xFF & vm->arena[mdx(cmd->idx + 6)].acb;
 	//res = cmd->reg[pc - 1] + ((arg1 + arg2) % IDX_MOD);
-	res = (arg1 + arg2);
+	res = ((short)arg1 + (short)arg2);
 	//ft_printf("sti arg1|%d| arg2|%d| %d\n", arg1, arg2, res);
 	return (res);
 }
 
 int		vm_rir_sti(t_vm *vm, t_cmd *cmd)
 {
-	int		res;
-	int		arg1;
-	short	arg2;
-	short	dir;
+	int				res;
+	int				arg1;
+	unsigned short	arg2;
+	short			dir;
 	
 	res = 0;
 	dir = 0xFF & vm->arena[mdx(cmd->idx + 3)].acb;
 	dir <<= 8;
 	dir += 0xFF & vm->arena[mdx(cmd->idx + 4)].acb;
-	dir = (dir % IDX_MOD);
+	dir = ((short)dir % IDX_MOD);
 	arg1 = 0xFF & vm->arena[mdx(cmd->idx + dir)].acb;
 	arg1 <<= 8;
 	arg1 += 0xFF & vm->arena[mdx(cmd->idx + dir + 1)].acb;
@@ -81,9 +81,9 @@ int		vm_rir_sti(t_vm *vm, t_cmd *cmd)
 
 int		vm_rrd_sti(t_vm *vm, t_cmd *cmd)
 {
-	int		res;
-	int		arg1;
-	short	arg2;
+	int				res;
+	int				arg1;
+	unsigned short	arg2;
 	
 	res = 0;
 	arg1 = 0xFF & vm->arena[mdx(cmd->idx + 3)].acb;
@@ -91,15 +91,15 @@ int		vm_rrd_sti(t_vm *vm, t_cmd *cmd)
 	arg2 = 0xFF & vm->arena[mdx(cmd->idx + 4)].acb;
 	arg2 <<= 8;
 	arg2 += 0xFF & vm->arena[mdx(cmd->idx + 5)].acb;
-	res = (arg1 + arg2);
+	res = (arg1 + (short)arg2);
 	return (res);
 }
 
 int		vm_rrr_sti(t_vm *vm, t_cmd *cmd)
 {
 	int		res;
-	short	arg1;
-	short	arg2;
+	int	arg1;
+	int	arg2;
 	
 	res = 0;
 	arg1 = 0xFF & vm->arena[mdx(cmd->idx + 3)].acb;
@@ -110,10 +110,10 @@ int		vm_rrr_sti(t_vm *vm, t_cmd *cmd)
 
 int		vm_rid_sti(t_vm *vm, t_cmd *cmd)
 {
-	int		res;
-	short	arg1;
-	short	dir;
-	int		arg2;
+	int				res;
+	unsigned short	arg1;
+	short			dir;
+	int				arg2;
 	
 	res = 0;
 	arg1 = 0xFF & vm->arena[mdx(cmd->idx + 3)].acb;
@@ -131,7 +131,7 @@ int		vm_rid_sti(t_vm *vm, t_cmd *cmd)
 	arg2 += 0xFF & vm->arena[mdx(cmd->idx + dir + 2)].acb;
 	arg2 <<= 8;
 	arg2 += 0xFF & vm->arena[mdx(cmd->idx + dir + 3)].acb;
-	res = (arg1 + arg2);
+	res = ((short)arg1 + arg2);
 	return (res);
 }
